Moved defvjp convenience wrappers to their own module, `extend`, which also imports Box, VSpace and primitive. `extend` now has everything needed for wrapping a numerical library, which should make the API clearer.
