+++ autograd/convenience_wrappers.py
@@ -6,8 +6,8 @@ from inspect import getargspec
 import warnings
 
 from .wrap_util import unary_to_nary
-from .tracer import primitive
-from .core import make_vjp as _make_vjp, make_jvp as _make_jvp, defvjp_argnum, vspace
+from .core import make_vjp as _make_vjp, make_jvp as _make_jvp
+from .extend import primitive, defvjp_argnum, vspace
 
 import autograd.numpy as np
