+++ autograd/scipy/stats/multivariate_normal.py
@@ -2,9 +2,9 @@ from __future__ import absolute_import
 import scipy.stats
 
 import autograd.numpy as np
-from autograd.core import primitive
+from autograd.extend import primitive
 from autograd.numpy.numpy_vjps import unbroadcast_f
-from autograd.core import primitive, defvjp, defvjps, defvjp_is_zero, defvjp_argnum
+from autograd.extend import primitive, defvjp, defvjps, defvjp_is_zero, defvjp_argnum
 
 
 pdf    =  primitive(scipy.stats.multivariate_normal.pdf)
